refactor: move tag removal logic to separate file
