+++ main.py
@@ -1,5 +1,5 @@
 """
-MC Structure cleaner
+MC Structure Cleaner
 By: Nyveon and DemonInTheCloset
 Thanks: lleheny0
 Special thanks: panchito, tomimi, puntito,
@@ -7,158 +7,22 @@ and everyone who has reported bugs and given feedback.
 
 Modded structure cleaner for minecraft. Removes all references to non-existent
 structures to allow for clean error logs and chunk saving.
+
+Project structure:
+    main.py - Command Line Interface
+    remove_tags.py - Main logic
+    tests - Unit tests
 """
 
-import time
-import os
-import anvil
-import itertools as it
 from argparse import ArgumentParser, Namespace
 from pathlib import Path
-from multiprocessing import Pool, cpu_count
-from typing import Set, Tuple
+from remove_tags import remove_tags
+from multiprocessing import cpu_count
 
-# --- Constants ---
 
 VERSION = "1.6"
-
-VANILLA_STRUCTURES = {
-    "bastion_remnant",
-    "buried_treasure",
-    "desert_pyramid",
-    "endcity",
-    "fortress",
-    "igloo",
-    "jungle_pyramid",
-    "mansion",
-    "mineshaft",
-    "monument",
-    "nether_fossil",
-    "ocean_ruin",
-    "pillager_outpost",
-    "ruined_portal",
-    "shipwreck",
-    "stronghold",
-    "swamp_hut",
-    "village",
-}
-
-# Data version difference between 1.17.1 and 1.18 snapshots+
-NEW_DATA_VERSION = 2800
-
-
-# Print separator
-def sep():
-    """Print separator line"""
-    print("----------------------------------")
-
-
-# Removing Tags
-def _remove_tags_region(args: Tuple[Set[str], Path, Path, str]) -> int:
-    """Wrapper for removing tags from a region file"""
-    return remove_tags_region(*args)
-
-
-def remove_tags_region(to_replace: Set[str],
-                       src: Path, dst: Path, mode: str) -> int:
-    """Remove tags in to_replace from the src region
-    Write changes to dst/src.name"""
-    start: float = time.perf_counter()
-    count: int = 0
-
-    print("Checking file:", src)
-
-    # Check if it's even an .mca file
-    print(src)
-    if len(str(src)) > 4:
-        if str(src)[-1:-5:-1] != "acm.":
-            print(f"{src} is not a valid region file.")
-            return 0
-    else:
-        print(f"{src} is not a valid file.")
-        return 0
-
-    # Check if file isn't empty
-    if os.path.getsize(src) == 0:
-        print(f"{src} is empty.")
-        return 0
-
-    coords = src.name.split(".")
-    region = anvil.Region.from_file(str(src.resolve()))
-    new_region = anvil.EmptyRegion(int(coords[1]), int(coords[2]))
-    removed_tags = set()
-
-    # Lambda function for checking if a tag is valid
-    if mode == "purge":
-        def check_tag(_tag):
-            return _tag.name.lower() not in VANILLA_STRUCTURES
-    else:
-        def check_tag(_tag):
-            return _tag.name in to_replace
-
-    # Check chunks
-    for chunk_x, chunk_z in it.product(range(32), repeat=2):
-        # Chunk Exists
-        if region.chunk_location(chunk_x, chunk_z) != (0, 0):
-            data = region.chunk_data(chunk_x, chunk_z)
-            data_copy = region.chunk_data(chunk_x, chunk_z)
-
-            if hasattr(data["Level"]["Structures"]["Starts"], 'tags'):
-                for tag in data["Level"]["Structures"]["Starts"].tags:
-                    if check_tag(tag):
-                        del data_copy["Level"][
-                            "Structures"]["Starts"][tag.name]
-                        count += 1
-                        removed_tags.add(tag.name)
-
-            if hasattr(data["Level"]["Structures"]["References"], 'tags'):
-                for tag in data["Level"]["Structures"]["References"].tags:
-                    if check_tag(tag):
-                        del data_copy["Level"][
-                            "Structures"]["References"][tag.name]
-                        count += 1
-                        removed_tags.add(tag.name)
-
-            # Add the modified chunk data to the new region
-            new_region.add_chunk(anvil.Chunk(data_copy))
-
-    # Save Region
-    new_region.save(str((dst / src.name).resolve()))
-
-    end: float = time.perf_counter()
-    print(f"File {src}: {count} \
-        instances of tags removed in {end - start:.3f} s")
-
-    # Output for purge mode (removed non vanilla tags per file)
-    if mode == "purge" and len(removed_tags) != 0:
-        print("Non-vanilla tags found:")
-        print(removed_tags)
-        sep()
-
-    return count
-
-
-def remove_tags(tags: Set[str],
-                src: Path, dst: Path, jobs: int, mode: str) -> None:
-    """Removes tags from src region files and writes them to dst"""
-    with Pool(processes=jobs) as pool:
-        start = time.perf_counter()
-
-        data = zip(it.repeat(tags),
-                   src.iterdir(),
-                   it.repeat(dst),
-                   it.repeat(mode))
-        count = sum(pool.map(_remove_tags_region, data))
-
-        end = time.perf_counter()
-
-        sep()
-        if mode == "purge":
-            print(f"Done!\nRemoved {count} instances of non-vanilla tags")
-        else:
-            print(f"Done!\nRemoved {count} instances of tags: {tags}")
-
-        print(f"Took {end - start:.3f} seconds")
+NEW_DATA_VERSION = 2800  # 1.17.1 vs 1.18
+SEP = "-" * 32
 
 
 # Environment
@@ -167,7 +31,7 @@ def setup_environment(new_region: Path) -> bool:
     if new_region.exists():
         print(f"{new_region.resolve()} exists, this may cause problems")
         proceed = input("Do you want to proceed regardless? [y/N] ")
-        sep()
+        print(SEP)
         return proceed.startswith("y")
 
     new_region.mkdir()
@@ -218,31 +82,33 @@ def get_args() -> Namespace:
     return parser.parse_args()
 
 
+def process_args(args: Namespace):
+    """Process CLI Arguments"""
+    processed = []
+    processed.append(set(args.tag))
+    processed.append(Path(f"{args.output}/new_region{args.region}"))
+    processed.append(Path(f"{args.path}/{args.region}/region"))
+    processed.append(args.jobs)
+    return processed
+
+
 def main() -> None:
-    args = get_args()
+    args = process_args(get_args())
+    to_replace, new_region, world_region, num_processes = args
 
-    to_replace = set(args.tag)
-    new_region = Path(f"{args.output}/new_region{args.region}")
-    world_region = Path(f"{args.path}/{args.region}/region")
-    num_processes = args.jobs
-
-    sep()
+    print(SEP)
 
     # Force purge mode if no tag is given, otherwise normal.
-    mode = "normal"
-    if args.tag == "":
+    mode = "purge" if not to_replace else "normal"
+    if mode == "purge":
         print("No tag given, will run in purge mode.")
-        mode = "purge"
-    else:
-        print("Tag(s) given, will run in normal mode.")
-
-    # Feedback as to what the program is about to do.
-    if mode == "normal":
-        print(f"Replacing {to_replace} in all region files in {world_region}.")
-    elif mode == "purge":
         print(f"Replacing all non-vanilla structures in \
             all region files in {world_region}.")
-    sep()
+    else:
+        print("Tag(s) given, will run in normal mode.")
+        print(f"Replacing {to_replace} in all region files in {world_region}.")
+
+    print(SEP)
 
     # Check if world exists
     if not world_region.exists():
@@ -255,16 +121,13 @@ def main() -> None:
         return None
 
     n_to_process = len(list(world_region.iterdir()))
-
     remove_tags(to_replace, world_region, new_region, num_processes, mode)
 
     # End output
-    sep()
-    print(f"Processed {n_to_process} files")
+    print(f"{SEP}\nProcessed {n_to_process} files")
     print(f"You can now replace {world_region} with {new_region}")
     return None
 
 
-# When running
 if __name__ == "__main__":
     main()
