+++ tests/test_remove_tags.py
@@ -1,0 +1,4 @@
+"""
+MC Structure Cleaner
+Tests tag removal logic
+"""
